Split editor playback selection warmup
