+++ apps/frontend/src/components/editor/editorPlaybackWarmupTypes.ts
@@ -1,0 +1,14 @@
+export type RefValue<T> = {
+  current: T;
+};
+
+export interface PlaybackReadyRange {
+  startTime: number;
+  endTime: number;
+  readyUntilTime: number;
+  status: "idle" | "warming" | "ready";
+}
+
+export interface WarmClipSelectionOptions {
+  extendToSelectionEnd?: boolean;
+}
